修复一些类型转换的问题

+++ kernel/syscall/arg.h
@@ -1,8 +1,8 @@
 #ifndef ARG_H
 #define ARG_H
 #include "../include/stdint.h"
-int argstr(int n, char** pp);
-int argptr(int n,char** pp,int size);
-int argint(int n, uint64_t* ip);
+int64_t argstr(int n, char** pp);
+int64_t argptr(int n,char** pp,int size);
+int64_t argint(int n, uint64_t* ip);
 
 #endif /* ARG_H */
